Set PCRs solver to ParallelCaseSolver.
ParallelCaseRunner (PCR) had it's solver set to normal CaseSolver, so it
didn't use multiple cores. I have written a minimal constructor to
change this. Also `n_cores` couldn't be read from config file, so it is
a keyword argument of PCRs constructor now. It is better, because every
user would want a different setting.